debog 2-1

+++ advertise/views.py
@@ -7,7 +7,7 @@ from rest_framework import status, permissions
 from rest_framework.parsers import JSONParser, MultiPartParser, FormParser
 from rest_framework.decorators import api_view, permission_classes, parser_classes
 from rest_framework.permissions import IsAuthenticated, AllowAny
-from rest_framework.exceptions import MethodNotAllowed, NotAcceptable, NotAcceptable
+from rest_framework.exceptions import MethodNotAllowed, NotAcceptable, NotAcceptable, NotFound
 from .models import Packet, Travel, Offer, Bookmark, Report, PacketPicture
 from account.models import User, Country, City
 from .serializers import *
@@ -201,15 +201,25 @@ def visit_travel(request, pk):
 
 
 @permission_classes([IsAuthenticated])
-@api_view(['DELETE'])
+@api_view(['DELETE', 'GET'])
 def bookmark(request, slug):
     user = User.objects.get(pk=request.user.id)
     packet = Packet.objects.get(slug=slug)
-    bookmark = Bookmark.objects.filter(owner=user, packet=packet)
-    bookmark.delete()
-    return HttpResponse(status=204)
-
-
+    if request.method == 'GET':
+        try:
+            bookmark = Bookmark.objects.get(owner=user, packet=packet)
+            return JsonResponse({"bookmark":True})
+        except Bookmark.DoesNotExist:
+            return JsonResponse({"bookmark":False})
+    if request.method == 'DELETE':
+        try:
+            bookmark = Bookmark.objects.get(owner=user, packet=packet)
+            bookmark.delete()
+            return HttpResponse(status=204)
+        except Bookmark.DoesNotExist:
+            raise NotFound(detail="آگهی مورد نظر پیدا نشد")
+        
+        
 @permission_classes([IsAuthenticated])
 @api_view(['GET','POST'])
 def bookmark_list(request):
@@ -220,8 +230,8 @@ def bookmark_list(request):
         return JsonResponse(serializer.data, safe=False)
     elif request.method == 'POST':
         packet = Packet.objects.get(slug=request.data.get('packet'))
-        if packet.owner == user :
-            count = Bookmark.objects.filter(owner=user).exclude(packet=packet).count()
+        if packet.owner != user :
+            count = Bookmark.objects.filter(owner=user, packet=packet).count()
             if count == 0 :
                 data = {
                     "packet": packet.id
@@ -234,8 +244,9 @@ def bookmark_list(request):
             else:
                 raise NotAcceptable(detail="قبلا نشان شده است")
         else:
-            detail = "این آگهی برای خودتان است. امکان ثبت پیشنهاد وجود ندارد"
-            return JsonResponse(str(detail), status=401, safe=False)
+            detail = "! این آگهی برای خودتان است"
+            raise NotAcceptable(detail)
+
 
 
 @permission_classes([IsAuthenticated])
@@ -266,21 +277,27 @@ def offer(request):
     slug = request.data.get("slug")
     user = User.objects.get(pk=request.user.id)
     packet = Packet.objects.get(slug=request.data.get("packet"))
-    if packet.owner != user :
-        price = request.data.get("price")
-        description = request.data.get("description")
-        travel_slug = request.data.get("travel")
-        travel = Travel.objects.get(slug=travel_slug)
-        data = request.data
-        serializer = OfferDeserializer(data=data)
-        if serializer.is_valid():
-            serializer.save(packet=packet, travel=travel)
-            return JsonResponse(serializer.data, status=201)
-        return JsonResponse(serializer.errors, status=400)
+    travel = Travel.objects.get(slug=request.data.get("travel"))
+    offer = Offer.objects.filter(travel=travel, packet=packet)
+    print(offer)
+    if offer.count == 0 :
+        if packet.owner != user :
+            price = request.data.get("price")
+            description = request.data.get("description")
+            travel_slug = request.data.get("travel")
+            travel = Travel.objects.get(slug=travel_slug)
+            data = request.data
+            serializer = OfferDeserializer(data=data)
+            if serializer.is_valid():
+                serializer.save(packet=packet, travel=travel)
+                return JsonResponse(serializer.data, status=201)
+            return JsonResponse(serializer.errors, status=400)
+        else:
+            detail = "این آگهی برای خودتان است. امکان ثبت پیشنهاد وجود ندارد"
+            return JsonResponse(str(detail), status=401, safe=False)
     else:
-        detail = "این آگهی برای خودتان است. امکان ثبت پیشنهاد وجود ندارد"
-        return JsonResponse(str(detail), status=401, safe=False)
-
+        raise NotAcceptable(detail="برای هر آگهی بیش از یک پیشنهاد مجاز نمی‌باشد")
+    
   
 @permission_classes([IsAuthenticated]) # TODO is need owner of object
 @api_view(['POST'])
@@ -313,7 +330,7 @@ def get_picture(request, pk):
 @permission_classes([IsAuthenticated])        
 def get_user_offer(request):
     user = User.objects.get(pk=request.user.id)
-    offer = Offer.objects.filter(travel__owner=user).exclude(status=8)
+    offer = Offer.objects.filter(travel__owner=user).exclude(status=8).order_by('-updated_at')
     serializer = OfferSerializer(offer, many=True)
     return JsonResponse(serializer.data, safe=False)
